+++ src/server/client.hpp
@@ -7,8 +7,8 @@
 
 //cette structure represente les client dans la liste des clients connécter au serveur
 struct Client{
-    long long projectId = 0;
-    std::atomic<long long> id = 0;
+    uint projectId = 0;
+    std::atomic<uint> id = 0;
     std::shared_ptr<sf::TcpSocket> sock;
     sf::IpAddress getAddress();
 };
